+++ sorgenti/lib/svd_cuda_engine.cu
@@ -16,14 +16,14 @@ void SvdCudaEngine::init(Matrix* matrix){
     cudaMalloc ((void**)&deviceInfo, sizeof(int));
     
     //Save matrix mem dimension
-    size_t space = (matrix->ld)*(matrix->n)*sizeof(double);
+    size_t space = (matrix->ld)*(matrix->n)*sizeof(float);
 
     //Allocate memory on device
     //std::cout<<"risultato cudaMalloc di A: "<<
     cudaMalloc((void**) &deviceA, space);//<<std::endl;
-    cudaMalloc((void**) &deviceU, (matrix->ld)*(matrix->m)*sizeof(double));
-    cudaMalloc((void**) &deviceS, (matrix->n)*sizeof(double));
-    cudaMalloc((void**) &deviceVT, (matrix->n)*(matrix->n)*sizeof(double));
+    cudaMalloc((void**) &deviceU, (matrix->ld)*(matrix->m)*sizeof(float));
+    cudaMalloc((void**) &deviceS, (matrix->n)*sizeof(float));
+    cudaMalloc((void**) &deviceVT, (matrix->n)*(matrix->n)*sizeof(float));
 
     //Copy matrix on device
     //std::cout<<"risultato memcpy di A: "<<
@@ -33,13 +33,13 @@ void SvdCudaEngine::init(Matrix* matrix){
 
 std::vector<Matrix*> SvdCudaEngine::getOutputMatrices(){
 
-    double *hostU, *hostVT, *hostS;
+    float *hostU, *hostVT, *hostS;
     Matrix *outputU, *outputVT, *outputS;
 
     //Cpu matrix resource allocation
-    hostU = new double[(input->m)*(input->m)]();
-    hostVT = new double[(input->n)*(input->n)]();
-    hostS = new double[input->n]();
+    hostU = new float[(input->m)*(input->m)]();
+    hostVT = new float[(input->n)*(input->n)]();
+    hostS = new float[input->n]();
 
     //Allocate memory on host
     outputU = new Matrix(input->ld, input->m, input->m, hostU);
@@ -47,9 +47,9 @@ std::vector<Matrix*> SvdCudaEngine::getOutputMatrices(){
     outputS = new Matrix (1, input->n, 1, hostS);
 
     //Copy back to host
-    cudaMemcpy(hostU, deviceU, (outputU->ld)*(outputU->m)*sizeof(double), cudaMemcpyDeviceToHost );
-    cudaMemcpy(hostVT, deviceVT, (outputVT->n)*(outputVT->n)*sizeof(double), cudaMemcpyDeviceToHost );
-    cudaMemcpy(hostS, deviceS, (outputS->n)*sizeof(double), cudaMemcpyDeviceToHost);
+    cudaMemcpy(hostU, deviceU, (outputU->ld)*(outputU->m)*sizeof(float), cudaMemcpyDeviceToHost );
+    cudaMemcpy(hostVT, deviceVT, (outputVT->n)*(outputVT->n)*sizeof(float), cudaMemcpyDeviceToHost );
+    cudaMemcpy(hostS, deviceS, (outputS->n)*sizeof(float), cudaMemcpyDeviceToHost);
 
     //Save SVD
     output = {outputU, outputS, outputVT};
